sec manager, move connection specific data out of server<>
Move mtu and connection encryption state to its own class.
So as to have them available independently from server<>

They can now be used by other classes without knowing the server
type.
